Fixed a number of node types (while, until, block_pass, cref, memo); nodewrap was dumping the wrong members for these types.
Fixed rval and noex members to behave more like the other members (they
are no longer special cases).  Node and object types are now dumped
and loaded using the same code (the objects flags are used to determine
type).  Noex is treated as an integer, same as argc/state/cnt.

Fixed nodeinfo.h.rpp to include prototype for define_node_elem_methods.

Cleaned up code for wrap_node.  Fixed bug that was causing random
segfaults (ref count should be initialized to 1, not 0, as there is at
least one reference to the node to begin with, the one we are
returning).  Fixed another bug that also caused segfaults (the node ids
we were storing were too big for Fixnums, so we were creating Bignum
objects while the GC was sweeping, which is not good).

Added unwrap_node function.  Cleaned up calling code to use it.

Check to see if nd_file is null before creating a string out of it.  If
it is null, then return Qnil.

Added Node#nd_line method.

Fixed proc_dump and proc_load to dump/load both the body and var nodes.
This ensures that the proc has the correct arity when it is reloaded.

Made UnboundProc a subclass of Proc (following the lead of UnboundMethod
and Method).  This allows methods that work with Proc to be used with
UnboundProc.  Disabled UnboundProc#call and UnboundProc#[].  Removed
mark_unbound_proc().  Added create_proc() function, which simplifies the
creation of Proc and UnboundProc objects.

Removed insert_sorted_modules_proc, since it was no longer being used.
Mark lookup_modules_proc as a global variable, otherwise random crashes
occur.

Pretty-print nodes using ascii-art tree notation.  Print file and line
number with each node that is printed.

Synchronize stdout and stderr when running the tests.  This is useful
when piping the test output through less.

Added proc node samples in addition to the method node samples.  This
allows us to test a wider variety of nodes (not all nodes can be
evaluated inside a method).