Updated code to work on Ruby 1.9 (cvs HEAD).   - Ruby 1.9 has no NODE_NEWLINE; don't try to use it.   - Added NODE_DSYM and NODE_ATTRASGN.   - Correctly handle flip states by using rb_is_local_id().   - Can now dump names of singleton classes (Ruby reports them as empty     string, but we dump them as OuterModule::InnerClass::<Singleton>.   - Use nd_type() to get the node's type instead of reproducing that     macro inline.   - Added _dump and _load for UnboundMethod, since UnboundMethod no     longer inherits from Method.   - Changed node sample for colon2 from Kernel::Object to     Object::Kernel, since Object can no longer be looked up from Kernel.
Fixed a bug in the wrap_node.  It's silly to try to use reference counts
when we have a garbage collector, and we get notified when the object
gets collected.  The reference count code was broken, so I removed it in
favor of simpler code.

Mark the class restorer as a global variable; it's not assigned to a
constant, so it's free to be GC'd.

Renamed Node#pretty_print to Node#tree and added Node#pretty_print as a
wrapper for Node#tree.

Added methods for generating a graph of a class hierarchy.

Added some entries to TODO.